requirements.txt
# Core data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# Database connectivity
clickhouse-connect>=0.6.0

# Environment configuration
python-dotenv>=0.19.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2022.1

# Performance optimization (optional)
numba>=0.56.0